requirements.txt
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.2
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nltk==3.3
psycopg2==2.7.5
pylint==2.1.1
python-decouple==3.1
pytz==2018.5
requests==2.19.1
six==1.11.0
textblob==0.15.1
typed-ast==1.1.0
urllib3==1.23
whitenoise==4.1
wrapt==1.10.11
